feat: 팀 단일 조회 기능 구현
- 팀 단일 조회 기능 구현
- FavTeamRepository 상속 추가(JpaFavTeamRepository)
- 타입 수정(TeamInfoResponse-favCnt)

+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/response/TeamInfoResponse.java
@@ -12,11 +12,11 @@ public record TeamInfoResponse(
     String profile,
     String homeLink,
     String snsLink,
-    Integer favCnt,
+    long favCnt,
     Boolean isFavorite
 ) {
 
-  public static TeamInfoResponse from(Team team, Integer favCnt, boolean isFav) {
+  public static TeamInfoResponse from(Team team, long favCnt, boolean isFav) {
     return new TeamInfoResponse(
         team.getTeamId(),
         team.getName(),
